status:   Both tasks completed:
  A) Forced Stall Test ✅

  Files created/modified:
  - src/workers/mock.ts - Mock worker with hang modes
  - src/supervisor/runner.ts - Added mock worker integration + stop-check in retry loop
  - test-fixtures/forced-stall/ - Test config and task

  Verified events in timeline:
  1. ✅ worker_call_timeout stop triggered after 12s cap exceeded
  2. ✅ auto_resume_scheduled with 100ms delay
  3. ✅ auto_resume_started resuming from PLAN phase
  4. ✅ auto_resume_count incremented (0 → 1)
  5. ✅ stop_reason cleared on resume

  Usage:
  AGENT_MOCK_WORKER=hang node dist/cli.js run --task ... --auto-resume

  B) agent status --all ✅

  Output:
  RUN ID          STATUS   PHASE    PROGRESS  AGE  STOP REASON          RESUMES  IN-FLIGHT
  ----------------------------------------------------------------------------------------
  20251227052512  running  PLAN     1/1       8m   -                    1        -
  20251227053059  stopped  STOPPED  1/1       2m   worker_call_timeout  1        -
  20251227052719  stopped  STOPPED  1/1       6m   worker_call_timeout  1        -

  Total: 6 runs (1 running, 5 stopped)

  Usage:
  agent status --all              # All runs
  agent status <runId>            # Single run (unchanged)

  All 232 tests pass.

+++ src/config/schema.ts
@@ -69,8 +69,8 @@ const resilienceSchema = z.object({
     .array(z.number().int().positive())
     .nonempty()
     .default([30000, 120000, 300000]), // 30s, 2min, 5min
-  /** Hard cap on worker call duration in minutes (kills hung workers) */
-  max_worker_call_minutes: z.number().int().positive().default(45)
+  /** Hard cap on worker call duration in minutes (kills hung workers). Supports decimals for testing. */
+  max_worker_call_minutes: z.number().positive().default(45)
 });
 
 export const agentConfigSchema = z.object({
